pyproject.toml
[tool.poetry]
name = "provenaclient"
version = "0.6.0"
description = "A client library for interfacing with a Provena instance."
authors = ["RRAP <rrapisdev@gmail.com>"]
license = "None"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.27.0"
pydantic = "<2.0"
python-jose = "<3.3.0"
provena-interfaces = "^2.6.0"
botocore = "1.30.1"
cloudpathlib = { extras = ["s3"], version = "0.15.1" }

[tool.poetry.dev-dependencies]
mypy = "<1.9.0"
pytest = "<8.1.1"
python-semantic-release = "<9.4.0"
pytest-cov = "<5.0.0"
myst-nb = "^1.0.0"
sphinx-autoapi = "^3.0.0"
sphinx-rtd-theme = "<2.0.0"
types-requests="*"

[tool.semantic_release]
version_variable = [
    "src/provenaclient/__init__.py:__version__",
    "pyproject.toml:version"
]                                                   # version location
branch = "main"                                      # branch to make releases of
changelog_file = "CHANGELOG.md"                      # changelog file
build_command = "pip install poetry && poetry install && poetry build"
dist_path = "dist/"
upload_to_release = true
upload_to_pypi = true 
remove_dist = false
patch_without_tag = true


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
